requirements.txt
Flask==2.0.1
Flask-SocketIO==5.0.1
requests==2.25.1
pytest==6.2.4
pytest-flask==1.2.0
jsonschema==3.2.0
python-socketio==5.0.4
markupsafe==2.0.1